[#174] Refactor login.js to enable Login by default
Leave the Login button enabled on page load, so that it's clickable when there are auto-filled values.
Validate the fields on form submission, so that clicking the Login button with empty fields will show an error state instead of submitting the form, as will pressing enter.
Don't show a success state on the fields, since it seems weird to claim success before the login happens.  They just need to be non-empty.
Rely on the HTML5 email validation message instead of checking for @ in the email field.
Use one event handler for all input events.