Fix button layout in delivery confirmation dialogs
Fix #184.
Add padding to the alert and make the buttons the same width.
Make the alert z-index 1000 so it floats above the header as well.  Currently, you can log out while the dialog is open.
Add the same shadow as the header to the dialog, so there's a little separation with whatever's underneath.
Update the text of the cancel ringdown and redirect confirmation dialogs.
Use standard units for the button padding and margins.
Fix the missing icon on the delivery canceled dialog.