Fix modal overlay layout in all cases
Add .minh-100vh class to the ER container so it stretches the full viewport height even on the hospital tab.  Without that, the overlay contents get clipped on the bottom.
Switch Header back to position: sticky.
Switch UnconfirmedRingdowns back to position: absolute and make it 100vh high, so it doesn't stick to the bottom of the content.