Add support for outputting text with each screenshot
Write the steps to the metadata.json files.
Add screenshot() as a function to call in a script step.
Clear the output paths one by one, instead of all of them, so everything doesn't get wiped if we just build some guides.
Go to /logout before trying to login in the scripts.
Add login steps for a hospital user.
Clean up upload.js.