Pull default ringdown values from query params
This makes it easy to fill out the ringdown form for testing.
Add a defaultPayload prop to RingdownForm, and use it when creating the ringdown state.
Add a check for query params in EMS.
Move the code to finish creating the Ringdown model class into an IIFE, so it's more clearly util code.
Add valueFromString() method to FieldMetadata, to make it easy to convert a string to a correctly typed value for that field.